requirements.txt
opencv-python==4.11.0.86
numpy>=1.26.0
Pillow==10.0.0 